small optimisations for filter integration test
